Move the Config derive into its own module
